TASK: Refactor tooltip into its own component
