+++ src/components/pages/Expeditions/CreationDialog/SeedInput.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
 
-import P from 'components/atoms/P'
-
 import FormControl from '../FormControl'
 import FormLabel from './FormLabel'
 import Input from '../Input'
+import Tooltip from 'components/molecules/Tooltip'
 
 type Props = {
   seed: string
@@ -16,6 +15,15 @@ const SeedInput = ({ seed, handleInputChange }: Props) => {
     <FormControl component={'fieldset' as 'div'}>
       <FormLabel>Seed Settings</FormLabel>
 
+      <Tooltip>
+        The seed makes each expeditions rolls completely deterministic. By
+        importing/copying an existing expedition and re-using its seed each roll
+        will be exactly the same as in the original provided that all parameters
+        are unchanged and you perform exactly the same actions. That way you can
+        try to beat your highscore or let other players enjoy a particular
+        expedition by sharing your configuration.
+      </Tooltip>
+
       <Input
         id="expeditionSeed"
         label="Seed"
@@ -26,15 +34,6 @@ const SeedInput = ({ seed, handleInputChange }: Props) => {
         variant="outlined"
         data-test="input-seed"
       />
-
-      <P variant="subtitle2">
-        The seed makes each expeditions rolls completely deterministic. By
-        importing/copying an existing expedition and re-using its seed each roll
-        will be exactly the same as in the original provided that all parameters
-        are unchanged and you perform exactly the same actions. That way you can
-        try to beat your highscore or let other players enjoy a particular
-        expedition by sharing your configuration.
-      </P>
     </FormControl>
   )
 }
